Initial default recipe
This pulls in some logic from the various places we're already using docker. The
upstream cookbook doesn't have a default recipe and is only resource driven, so
this cookbook will make an assumption that we want a default docker service
running utilizing dynamic attributes for setting the various parameters for each
resource. A few other notable features:

- Prefer using upstream vendor repositories instead of distribution repos to
  ensure we're using the latest and greatest.
- This uses apt pinning and yum-versionlock to ensure we have the exact version
  of docker installed we want. That way we don't upgrade when we don't want to
  accidentally.