Add ppc64le support / Update to latest Docker-ce version
Upstream docker doesn't officially have packages for docker on ppc64le. This
uses a repository from the Unicamp university which provides some newer releases
than what's included in CentOS proper. This adds some logic to work around
issues between the two architectures.

In addition, this upgrades the docker version from 1.13.1 to 17.06.1 on CentOS
and 17.05.0 on Debian. There doesn't seem to be a release of 17.06.1 on Debian
yet so we had to go with the latest version on there. This will upgrade docker
on the workstations so we'll have to watch and see what issues that cause.