Add workstation which deprecates workstation::docker
This moves the workstation::docker recipe into the osl-docker cookbook so that
it can be better tracked and tested.

The major changes that this move provides the following:

Switching from using the docker socket to using the docker tcp port. This should
remove the need to have the docker user in a workstation group and allow any
user access to the docker running on the workstation.

Upgrading from docker-1.10.3 to a newer more supported version and matches what
we use elsewhere.