Install docker-compose binary
This adds a compose recipe which installs the docker-compose binary. Upstream
doesn't provide a binary for ppc64le, so I made one for our use. This is one
release behind since I originally built the ppc64le binary before 1.19.0 was
released. I'll work on pushing up the code for that soon in a fork and then try
and get proper support upstream. This also unifies how we install docker-compose
globally if we need it elsewhere.

This will replace the pip installs we do on the workstations and will require us
removing docker-compose from the workstation environment when merged.