Optional TLS support for docker daemons
This adds the ability to having a docker daemon listen on a TLS encrypted port
with strict TLS verification. This relies on creating an internal certificate
authority (CA) using easy-rsa and signing server and client certificates. This
follows the upstream documentation [1] on how to deploy TLS properly on the
docker daemon.

This feature is needed for deploying the s390x docker hosts since they are
remote and we need to lock them down. This might also be useful for other use
cases in the future that I'm pondering. I'll be creating separate documentation
on the docker CA using easy-rsa.

[1] https://docs.docker.com/engine/security/https/#create-a-ca-server-and-client-keys-with-openssl