Update to docker-18.06.1
This also changes how we install docker on ppc64le to using the upstream tarball
instead of pulling from ftp.unicamp.br. This ensures that we always use the
latest version and are not dependent on when the other site builds their rpms.
It also aligns to the same way we install it on s390x.

In addition, ensure we remove the old docker-ce package on ppc64le so that it
properly installs the new tarball image.