Tell docker_service not to try an install_method
We install docker using the installation resources directly instead of doing it
in the docker_service resource. This means the recipe tries to install docker
twice which is not good. While we should probably do it via the docker_service
this seems to work well as is. Unfortunately we have a problem on hosts such as
openpower8 which run a mainline kernel and likely confuses the auto installer in
docker_service. For now setting it to none will work around this problem.