Add support for nvidia-docker
This primarily installs the nvidia-docker2 package which pulls in all of the
necessary dependent packages to get nvidia-docker to work. In addition, this
also installs a version pinned version of the nvidia-driver. The version
currently locked is what's needed to test the openpower-gpu1 test host and can
be changed later to a newer version if needed.

Some other changes:
- Management of /etc/docker/daemon.json so that we can completely manage the
  runtime configuration and not rely on the nvidia-docker2 package.