Add client recipe
In some cases, we only want to have the docker cli installed and not the
service. One example of this is on Jenkins servers where we don't want the
docker daemon running. Instead we want to run them on worker nodes so we can
scale it better.

This just adds support within the default recipe via an attribute. For ease of
use, adding of the client recipe just sets this attribute so we can simply add
it in a run list where we need it.