Install latest docker and remove version pinning
We no longer need to worry about pinning a specific version due to the fact that
upstream now finally supports proper native packages for ppc64le and s390x. This
will ensure we get the latest version moving forward.

For now we'll leave parts of the pinning logic, but change it to removing those
pins.

In addition, we'll switch to using upstream repositories for ppc64le and s390x
instead of using our repositories now that they're available. Bump the version
of the docker cookbook as it requires a fix for s390x to work properly.

Signed-off-by: Lance Albertson <lance@osuosl.org>