Rename config property to config_files in osl_dockercompose
There appears to be some kind of issue using the property name 'config' so we're
renaming this to 'config_files' to work around it. I don't think this is a
reserved word, but there might be some conflicts with the docker cookbook helper
methods.

Also fix helper method osl_dockercompose_running? to support multiple docker
compose instances at the same time.

In addition, include an additional test with multiple files to ensure this works
in the future.

Signed-off-by: Lance Albertson <lance@osuosl.org>