Fix logic around package name
It was assumed if we weren't using the upstream repo, then this must be a debian
based system so use the distro package name instead.

However now that we're also doing this with ppc64le on RHEL, we need to update
the logic since this is using the upstream package name still.

Signed-off-by: Lance Albertson <lance@osuosl.org>